Fix Source Module names to singular form and API documentation bugfixes
