+++ src/bundles/curve/curves_webgl.ts
