+++ src/bundles/curve/types.ts
