+++ src/bundles/stereo_sound/types.ts
